requirements.txt
# install with `pip install -r`
pwntools
pycryptodome
rich
imageio
